- translations from old qt3 version readded: fr, it, sr (unmaintained) - comment widget now expands if the text inside is larger than viewport - archive scanning added (needs libtar, libbzip2, lib7zip) - recent catalog menu reversed (last used now on top) - config dialog size optimizions - compile instructions added (linux, mac, windows)
git-svn-id: svn://svn.code.sf.net/p/cdcat/svn/trunk@202 23e97676-e3e9-41ff-bed1-4a4288a26c68

+++ src/import.cpp
@@ -2300,15 +2300,15 @@ bool importWhereIsItXml::endElement( const QString&, const QString & tag, const 
                 float s;
                 int st;
 
-                if ( size > ( uint ) ( 1024 * 1024 * 1024 * 2 ) ) {
+                if ( size > ( float ) ( 1024 * 1024 * 1024 * 2 ) ) {
                     s = ( double ) size / ( double ) ( 1024 * 1024 * 1024 );
                     st = GBYTE;
                 }
-                if ( size > ( uint ) ( 1024 * 1024 ) ) {
+                if ( size > ( float ) ( 1024 * 1024 ) ) {
                     s = ( double ) size / ( double ) ( 1024 * 1024 );
                     st = MBYTE;
                 }
-                if ( size > ( uint ) 1024 ) {
+                if ( size > ( float ) 1024 ) {
                     s = ( double ) size / ( double ) 1024;
                     st = KBYTE;
                 } else {
